// src/api/generated/models/TimeClockCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Clock in
 */
export type TimeClockCreate = {
    staff_id: string;
    clock_in: string;
    location_id?: (string | null);
    notes?: (string | null);
    business_id: string;
};